+++ dev-docs/CHANGELOG.md
@@ -14,8 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Standard Library
 
-- Add `StateInit.hasSameBasechainAddress` function: PR [#2187](https://github.com/tact-lang/tact/pull/2187)
+- Add the `StateInit.hasSameBasechainAddress` function: PR [#2187](https://github.com/tact-lang/tact/pull/2187)
 - Add doc comments for functions and structs: PR [#2267](https://github.com/tact-lang/tact/pull/2267)
+- Add the `cashback` function: PR [#2241](https://github.com/tact-lang/tact/pull/2241)
 
 ### TypeScript wrappers
 
@@ -43,6 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [verytactical](https://github.com/verytactical)
 - [Shvetc Andrei](https://github.com/Shvandre)
 - [Maksim Lagus](https://github.com/Kaladin13)
+- [Novus Nota](https://github.com/novusnota)
 
 ## [1.6.0] - 2025-02-28